[Rename] : HomePage -> LandingPage
