+++ src/components/LandingPage/DescriptionBox.tsx
