+++ src/routes/index.tsx
@@ -5,7 +5,7 @@ import OnboardingLayout from '@/layouts/OnboardingLayout';
 import PublicLayout from '@/layouts/PublicLayout';
 import AuthPage from '@/pages/AuthPage';
 import DashboardPage from '@/pages/DashboardPage';
-import HomePage from '@/pages/HomePage';
+import LandingPage from '@/pages/LandingPage';
 import NotesPage from '@/pages/NotesPage';
 import OnboardingPage from '@/pages/OnboardingPage';
 import ProfilePage from '@/pages/ProfilePage';
@@ -17,7 +17,7 @@ export const router = createBrowserRouter([
     path: '/',
     element: <PublicLayout />,
     children: [
-      { index: true, element: <HomePage /> },
+      { index: true, element: <LandingPage /> },
       { path: 'auth', element: <AuthPage /> },
     ],
   },
